Shuffle TFT init code around, remove SPI mutex
Re-reading the docs for the Titano show that it exclusively uses a parallel connection for the TFT. Removed SPI mutex around Wifi/TFT calls. Shifted TFT init code into FreeRTOS task as it would not update the display with the init and update code split between pre- and post- FreeRTOS initialisation.