Purge the collector output line when shutting the pump off
When there is low solar input, the pump often cycles for just a few minutes before shutting off for 5-10 minutes. My installation has long pipes (>10m) to the collectors, and a reasonable amount of warm water cools back down in the line before the pump starts again. This change adds a "line purge" to the pump shutdown, running the pump at 100 percent for 60 seconds.